requirements.txt
Django==2.2.0
psycopg2==2.7.5
djangorestframework==3.11.0
